pyproject.toml
[project]
name = "real-time-ai-using-socket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "black>=25.1.0",
    "google-generativeai>=0.8.4",
    "prompt-toolkit>=3.0.50",
    "python-dotenv>=1.1.0",
    "python-socketio>=5.12.1",
]
